Task2/auth_service_app/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.2.4
python-dotenv==1.0.1
